Refactor displacements into Trajectory
