+++ src/calculate/__init__.py
@@ -1,4 +1,3 @@
-from .displacements import Displacements
 from .extras import calculate_all
 from .tracer import Tracer
 from .vibration import Vibration
